Extract calendar picker
