+++ x-pack/solutions/observability/plugins/slo/public/pages/slo_details/components/history/slo_details_history.tsx
@@ -5,12 +5,10 @@
  * 2.0.
  */
 import {
-  EuiButton,
   EuiFlexGroup,
   EuiFlexItem,
   EuiPanel,
   EuiSuperDatePicker,
-  EuiText,
   EuiTitle,
   OnTimeChangeProps,
 } from '@elastic/eui';
@@ -18,39 +16,22 @@ import DateMath from '@kbn/datemath';
 import { i18n } from '@kbn/i18n';
 import { SLOWithSummaryResponse } from '@kbn/slo-schema';
 import moment from 'moment';
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import { ErrorRateChart } from '../../../../components/slo/error_rate_chart';
 import { useKibana } from '../../../../hooks/use_kibana';
 import { toDuration } from '../../../../utils/slo/duration';
 import { TimeBounds } from '../../types';
 import { EventsChartPanel } from '../events_chart_panel/events_chart_panel';
 import { HistoricalDataCharts } from '../historical_data_charts';
+import { CalendarPeriodPicker } from './calendar_period_picker';
 
 export interface Props {
   slo: SLOWithSummaryResponse;
   isAutoRefreshing: boolean;
 }
 
-function getPeriodLabel(slo: SLOWithSummaryResponse, calendarPeriod: number): string {
-  const duration = toDuration(slo.timeWindow.duration);
-  const isWeeklyCalendarAligned = duration.unit === 'w';
-  const now = moment().utc();
-
-  const start = now
-    .clone()
-    .subtract(calendarPeriod, isWeeklyCalendarAligned ? 'week' : 'month')
-    .startOf(isWeeklyCalendarAligned ? 'isoWeek' : 'month');
-  const end = now
-    .clone()
-    .subtract(calendarPeriod, isWeeklyCalendarAligned ? 'week' : 'month')
-    .endOf(isWeeklyCalendarAligned ? 'isoWeek' : 'month');
-
-  return `${start.format('LL')} - ${end.format('LL')}`;
-}
-
 export function SloDetailsHistory({ slo, isAutoRefreshing }: Props) {
   const { uiSettings } = useKibana().services;
-  const [calendarPeriod, setCalendarPeriod] = useState<number>(0);
 
   // TODO: refactor
   const [range, setRange] = useState<TimeBounds>(() => {
@@ -71,20 +52,6 @@ export function SloDetailsHistory({ slo, isAutoRefreshing }: Props) {
     };
   });
 
-  useEffect(() => {
-    if (slo.timeWindow.type === 'calendarAligned') {
-      const now = moment();
-      const duration = toDuration(slo.timeWindow.duration);
-      const unit = duration.unit === 'w' ? 'isoWeek' : 'month';
-      const durationUnit = duration.unit === 'w' ? 'week' : 'month';
-
-      return setRange({
-        from: moment.utc(now).subtract(calendarPeriod, durationUnit).startOf(unit).toDate(),
-        to: moment.utc(now).subtract(calendarPeriod, durationUnit).endOf(unit).toDate(),
-      });
-    }
-  }, [calendarPeriod, slo.timeWindow]);
-
   const onBrushed = ({ from, to }: TimeBounds) => {
     setRange({ from, to });
   };
@@ -94,34 +61,12 @@ export function SloDetailsHistory({ slo, isAutoRefreshing }: Props) {
       <EuiFlexGroup justifyContent="flexEnd" direction="row" gutterSize="s">
         <EuiFlexItem grow css={{ maxWidth: 500 }}>
           {slo.timeWindow.type === 'calendarAligned' ? (
-            <EuiFlexGroup direction="row" justifyContent="flexEnd" alignItems="center">
-              <EuiButton
-                size="s"
-                data-test-subj="sloSloDetailsHistoryPreviousButton"
-                onClick={() => {
-                  setCalendarPeriod((curr) => curr + 1);
-                }}
-              >
-                {i18n.translate('xpack.slo.sloDetailsHistory.previousPeriodButtonLabel', {
-                  defaultMessage: 'Previous',
-                })}
-              </EuiButton>
-              <EuiText size="s">
-                <p>{getPeriodLabel(slo, calendarPeriod)}</p>
-              </EuiText>
-              <EuiButton
-                size="s"
-                data-test-subj="sloSloDetailsHistoryNextButton"
-                disabled={calendarPeriod <= 0}
-                onClick={() => {
-                  setCalendarPeriod((curr) => curr - 1);
-                }}
-              >
-                {i18n.translate('xpack.slo.sloDetailsHistory.nextPeriodButtonLabel', {
-                  defaultMessage: 'Next',
-                })}
-              </EuiButton>
-            </EuiFlexGroup>
+            <CalendarPeriodPicker
+              slo={slo}
+              onChange={(updatedRange: TimeBounds) => {
+                setRange(updatedRange);
+              }}
+            />
           ) : (
             <EuiSuperDatePicker
               isLoading={false}
